requirements.txt
streamlit
pdfminer.six
scikit-learn
sentence-transformers
spacy
numpy
pandas
